+++ NEWS.md
@@ -5,6 +5,7 @@
 * **Rename** `trakt.getSeasons`  -> `trakt.seasons.summary` for consistency with the trakt API
 * **Rename** `trakt.show.season` -> `trakt.seasons.season` for consistency with the trakt API
 * **Rename** `getNameFromUrl`    -> `parse_trakt_url` to be more descriptive. Also, de-camelCasezation. 
+* Update vignette to reflext the above changes
 
 ### v0.10.3